test: split error recovery HMR test into multiple files
